restructure test utils
