+++ test-resources/tests/api-tests/dequeue/events.test.ts
@@ -5,7 +5,7 @@ import {
   EVENTS_API_INSTANCE,
   getActiveSessionId,
   pollForEvents,
-} from "../../testUtils/apiTestHelpers";
+} from "../utils/testFunctions";
 
 const ONE_SECOND = 1000;
 jest.setTimeout(45 * ONE_SECOND);
