+++ test-resources/tests/api-tests/sts-mock/jwks.test.ts
@@ -1,4 +1,4 @@
-import { STS_MOCK_API_INSTANCE } from "../../testUtils/apiTestHelpers";
+import { STS_MOCK_API_INSTANCE } from "../utils/testFunctions";
 
 describe("GET /.well-known/jwks.json", () => {
   it("returns 200 status code and the Json Web Key Set", async () => {
